requirements.txt
streamlit==1.17.0
pandas==1.2.4
requests==2.25.1
pillow==8.2.0